# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Kargo",
        "title": "Kargo API",
        "contact": {},
        "version": "v1alpha1"
    },
    "basePath": "/",
    "paths": {
...
        "/v1beta1/projects/{project}/analysis-runs/{analysis-run}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve an AnalysisRun resource from a project's namespace.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verifications",
                    "Project-Level"
                ],
                "summary": "Retrieve an AnalysisRun",
                "operationId": "GetAnalysisRun",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "AnalysisRun name",
                        "name": "analysis-run",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AnalysisRun custom resource (github.com/argoproj/argo-rollouts/pkg/apis/rollouts/v1alpha1.AnalysisRun)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/v1beta1/projects/{project}/analysis-templates/{analysis-template}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve an AnalysisTemplate resource from a project's\nnamespace.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verifications",
                    "Project-Level"
                ],
                "summary": "Retrieve an AnalysisTemplate",
                "operationId": "GetAnalysisTemplate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "AnalysisTemplate name",
                        "name": "analysis-template",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AnalysisTemplate custom resource (github.com/argoproj/argo-rollouts/pkg/apis/rollouts/v1alpha1.AnalysisTemplate)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an AnalysisTemplate resource from a project's namespace.",
                "tags": [
                    "Verifications",
                    "Project-Level"
                ],
                "summary": "Delete an AnalysisTemplate",
                "operationId": "DeleteAnalysisTemplate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "AnalysisTemplate name",
                        "name": "analysis-template",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                }
            }
        },
...
        "/v1beta1/projects/{project}/freight/{freight-name-or-alias}/alias": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Patch a Freight resource's human-friendly alias.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core",
                    "Project-Level"
                ],
                "summary": "Patch a Freight resource's alias",
                "operationId": "PatchFreightAlias",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Freight name or alias",
                        "name": "freight-name-or-alias",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "New alias",
                        "name": "newAlias",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/v1beta1/projects/{project}/stages/{stage}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a Stage resource from a project's namespace.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core",
                    "Project-Level"
                ],
                "summary": "Retrieve a Stage",
                "operationId": "GetStage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Stage name",
                        "name": "stage",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stage custom resource (github.com/akuity/kargo/api/v1alpha1.Stage)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a Stage resource from a project's namespace.",
                "tags": [
                    "Core",
                    "Project-Level"
                ],
                "summary": "Delete a Stage",
                "operationId": "DeleteStage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Stage name",
                        "name": "stage",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                }
            }
        },
...
        "/v1beta1/shared/cluster-analysis-templates/{cluster-analysis-template}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a ClusterAnalysisTemplate by name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verifications",
                    "Shared",
                    "Cluster-Scoped Resource"
                ],
                "summary": "Retrieve a ClusterAnalysisTemplate",
                "operationId": "GetClusterAnalysisTemplate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ClusterAnalysisTemplate name",
                        "name": "cluster-analysis-template",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ClusterAnalysisTemplate custom resource (github.com/argoproj/argo-rollouts/pkg/apis/rollouts/v1alpha1.ClusterAnalysisTemplate)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a ClusterAnalysisTemplate resource.",
                "tags": [
                    "Verifications",
                    "Shared",
                    "Cluster-Scoped Resource"
                ],
                "summary": "Delete a ClusterAnalysisTemplate",
                "operationId": "DeleteClusterAnalysisTemplate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ClusterAnalysisTemplate name",
                        "name": "cluster-analysis-template",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token authentication. Obtain token via OIDC/PKCE flow with your identity provider.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}